backend/app/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
python-dotenv==1.0.0
# psycopg2-binary==2.9.9
# google-auth==2.23.4
# redis==5.0.1
langchain
langchain-community
langchain-google-genai
langchain-huggingface
google-generativeai
faiss-cpu
sentence-transformers
pyinputplus
pypdf